eliminate compiler warnings

+++ src/transport/tcp_old/TCPConnectionUtil.cc
@@ -440,7 +440,7 @@ void TCPConnection::sendFin()
     tcpAlgorithm->ackSent();
 }
 
-void TCPConnection::sendSegment(int bytes)
+void TCPConnection::sendSegment(ulong bytes)
 {
     ulong buffered = sendQueue->getBytesAvailable(state->snd_nxt);
     if (bytes > buffered) // last segment?
